+++ spec/translate/translate.objectValue.spec.js
@@ -68,6 +68,8 @@ describe('accessing tree values', function() {
           translation: {                      
             test: { res: 'added __replace__',
                     id: 0,
+                    regex: /test/,
+                    func: function () {},
                     template: '4',
                     title: 'About...',
                     text: 'Site description',
@@ -88,6 +90,8 @@ describe('accessing tree values', function() {
         expect(i18n.t('test', { returnObjectTrees: true, replace: 'two' })).to.eql({ 
           res: 'added two',
           id: 0,
+          regex: resStore['en-US'].translation.test.regex,
+          func: resStore['en-US'].translation.test.func,
           template: '4',
           title: 'About...',
           text: 'Site description',