+++ src/i18next.translate.js
@@ -273,10 +273,11 @@ function _find(key, options) {
             x++;
         }
         if (value !== undefined) {
+            var valueType = Object.prototype.toString.apply(value);
             if (typeof value === 'string') {
                 value = applyReplacement(value, options);
                 value = applyReuse(value, options);
-            } else if (Object.prototype.toString.apply(value) === '[object Array]' && !o.returnObjectTrees && !options.returnObjectTrees) {
+            } else if (valueType === '[object Array]' && !o.returnObjectTrees && !options.returnObjectTrees) {
                 value = value.join('\n');
                 value = applyReplacement(value, options);
                 value = applyReuse(value, options);
@@ -291,8 +292,8 @@ function _find(key, options) {
                             'returned an object instead of string.';
                         f.log(value);
                     }
-                } else if (typeof value !== 'number') {
-                    var copy = {}; // apply child translation on a copy
+                } else if (valueType !== '[object Number]' && valueType !== '[object Function]' && valueType !== '[object RegExp]') {
+                    var copy = (valueType === '[object Array]') ? [] : {}; // apply child translation on a copy
                     f.each(value, function(m) {
                         copy[m] = _translate(ns + o.nsseparator + key + o.keyseparator + m, options);
                     });